+++ app/tfchain/tfchain.types.IO.js
@@ -1,5 +1,5 @@
 import {AssertionError, AttributeError, BaseException, DeprecationWarning, Exception, IndexError, IterableError, KeyError, NotImplementedError, RuntimeWarning, StopIteration, UserWarning, ValueError, Warning, __JsIterator__, __PyIterator__, __Terminal__, __add__, __and__, __call__, __class__, __envir__, __eq__, __floordiv__, __ge__, __get__, __getcm__, __getitem__, __getslice__, __getsm__, __gt__, __i__, __iadd__, __iand__, __idiv__, __ijsmod__, __ilshift__, __imatmul__, __imod__, __imul__, __in__, __init__, __ior__, __ipow__, __irshift__, __isub__, __ixor__, __jsUsePyNext__, __jsmod__, __k__, __kwargtrans__, __le__, __lshift__, __lt__, __matmul__, __mergefields__, __mergekwargtrans__, __mod__, __mul__, __ne__, __neg__, __nest__, __or__, __pow__, __pragma__, __proxy__, __pyUseJsNext__, __rshift__, __setitem__, __setproperty__, __setslice__, __sort__, __specialattrib__, __sub__, __super__, __t__, __terminal__, __truediv__, __withblock__, __xor__, abs, all, any, assert, bool, bytearray, bytes, callable, chr, copy, deepcopy, delattr, dict, dir, divmod, enumerate, filter, float, getattr, hasattr, input, int, isinstance, issubclass, len, list, map, max, min, object, ord, pow, print, property, py_TypeError, py_iter, py_metatype, py_next, py_reversed, py_typeof, range, repr, round, set, setattr, sorted, str, sum, tuple, zip} from './org.transcrypt.__runtime__.js';
-import {ConditionBaseClass, ConditionNil, UnlockHash, UnlockHashType} from './tfchain.types.ConditionTypes.js';
+import {ConditionBaseClass, ConditionLockTime, ConditionNil, ConditionUnlockHash, UnlockHash, UnlockHashType} from './tfchain.types.ConditionTypes.js';
 import {FulfillmentBaseClass, FulfillmentSingleSignature} from './tfchain.types.FulfillmentTypes.js';
 import * as FulfillmentTypes from './tfchain.types.FulfillmentTypes.js';
 import * as ConditionTypes from './tfchain.types.ConditionTypes.js';
@@ -315,7 +315,7 @@ Object.defineProperty (CoinInput, 'fulfillment', property.call (CoinInput, CoinI
 Object.defineProperty (CoinInput, 'parentid', property.call (CoinInput, CoinInput._get_parentid, CoinInput._set_parentid));;
 export var CoinOutput =  __class__ ('CoinOutput', [BaseDataTypeClass], {
 	__module__: __name__,
-	get __init__ () {return __get__ (this, function (self, value, condition, id, is_fee) {
+	get __init__ () {return __get__ (this, function (self, value, condition, id, is_fee, is_reward) {
 		if (typeof value == 'undefined' || (value != null && value.hasOwnProperty ("__kwargtrans__"))) {;
 			var value = null;
 		};
@@ -328,6 +328,9 @@ export var CoinOutput =  __class__ ('CoinOutput', [BaseDataTypeClass], {
 		if (typeof is_fee == 'undefined' || (is_fee != null && is_fee.hasOwnProperty ("__kwargtrans__"))) {;
 			var is_fee = false;
 		};
+		if (typeof is_reward == 'undefined' || (is_reward != null && is_reward.hasOwnProperty ("__kwargtrans__"))) {;
+			var is_reward = false;
+		};
 		if (arguments.length) {
 			var __ilastarg0__ = arguments.length - 1;
 			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
@@ -339,6 +342,7 @@ export var CoinOutput =  __class__ ('CoinOutput', [BaseDataTypeClass], {
 						case 'condition': var condition = __allkwargs0__ [__attrib0__]; break;
 						case 'id': var id = __allkwargs0__ [__attrib0__]; break;
 						case 'is_fee': var is_fee = __allkwargs0__ [__attrib0__]; break;
+						case 'is_reward': var is_reward = __allkwargs0__ [__attrib0__]; break;
 					}
 				}
 			}
@@ -356,6 +360,8 @@ export var CoinOutput =  __class__ ('CoinOutput', [BaseDataTypeClass], {
 		self.id = id;
 		self._is_fee = false;
 		self.is_fee = is_fee;
+		self._is_reward = false;
+		self.is_reward = is_reward;
 	});},
 	get from_json () {return __getcm__ (this, function (cls, obj) {
 		if (arguments.length) {
@@ -485,8 +491,8 @@ export var CoinOutput =  __class__ ('CoinOutput', [BaseDataTypeClass], {
 		}
 		else {
 		}
-		if (self._spendable_value == null) {
-			return Currency ();
+		if (self._spendable_value == null || self._spendable_value.__le__ (0)) {
+			return self.value;
 		}
 		return self._spendable_value;
 	});},
@@ -676,6 +682,44 @@ export var CoinOutput =  __class__ ('CoinOutput', [BaseDataTypeClass], {
 		}
 		self._is_fee = value;
 	});},
+	get _get_is_reward () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		return self._is_reward;
+	});},
+	get _set_is_reward () {return __get__ (this, function (self, value) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+						case 'value': var value = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (!(isinstance (value, bool))) {
+			var __except0__ = py_TypeError ('is is_reward is supposed to be a bool, cannot be {} ({})'.format (value, py_typeof (value)));
+			__except0__.__cause__ = null;
+			throw __except0__;
+		}
+		self._is_reward = value;
+	});},
 	get json () {return __get__ (this, function (self) {
 		if (arguments.length) {
 			var __ilastarg0__ = arguments.length - 1;
@@ -727,6 +771,7 @@ export var CoinOutput =  __class__ ('CoinOutput', [BaseDataTypeClass], {
 		encoder.add_all (self._value, self._condition);
 	});}
 });
+Object.defineProperty (CoinOutput, 'is_reward', property.call (CoinOutput, CoinOutput._get_is_reward, CoinOutput._set_is_reward));
 Object.defineProperty (CoinOutput, 'is_fee', property.call (CoinOutput, CoinOutput._get_is_fee, CoinOutput._set_is_fee));
 Object.defineProperty (CoinOutput, 'id', property.call (CoinOutput, CoinOutput._get_id, CoinOutput._set_id));
 Object.defineProperty (CoinOutput, 'condition', property.call (CoinOutput, CoinOutput._get_condition, CoinOutput._set_condition));
@@ -1249,15 +1294,21 @@ Object.defineProperty (BlockstakeOutput, 'condition', property.call (BlockstakeO
 Object.defineProperty (BlockstakeOutput, 'value', property.call (BlockstakeOutput, BlockstakeOutput._get_value, BlockstakeOutput._set_value));;
 export var MinerPayout =  __class__ ('MinerPayout', [object], {
 	__module__: __name__,
-	get __init__ () {return __get__ (this, function (self, value, unlockhash, id) {
+	get __init__ () {return __get__ (this, function (self, value, unlockhash, unlockheight, id, is_reward) {
 		if (typeof value == 'undefined' || (value != null && value.hasOwnProperty ("__kwargtrans__"))) {;
 			var value = null;
 		};
 		if (typeof unlockhash == 'undefined' || (unlockhash != null && unlockhash.hasOwnProperty ("__kwargtrans__"))) {;
 			var unlockhash = null;
 		};
+		if (typeof unlockheight == 'undefined' || (unlockheight != null && unlockheight.hasOwnProperty ("__kwargtrans__"))) {;
+			var unlockheight = null;
+		};
 		if (typeof id == 'undefined' || (id != null && id.hasOwnProperty ("__kwargtrans__"))) {;
 			var id = null;
+		};
+		if (typeof is_reward == 'undefined' || (is_reward != null && is_reward.hasOwnProperty ("__kwargtrans__"))) {;
+			var is_reward = true;
 		};
 		if (arguments.length) {
 			var __ilastarg0__ = arguments.length - 1;
@@ -1268,7 +1319,9 @@ export var MinerPayout =  __class__ ('MinerPayout', [object], {
 						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
 						case 'value': var value = __allkwargs0__ [__attrib0__]; break;
 						case 'unlockhash': var unlockhash = __allkwargs0__ [__attrib0__]; break;
+						case 'unlockheight': var unlockheight = __allkwargs0__ [__attrib0__]; break;
 						case 'id': var id = __allkwargs0__ [__attrib0__]; break;
+						case 'is_reward': var is_reward = __allkwargs0__ [__attrib0__]; break;
 					}
 				}
 			}
@@ -1279,8 +1332,12 @@ export var MinerPayout =  __class__ ('MinerPayout', [object], {
 		self.value = value;
 		self._unlockhash = null;
 		self.unlockhash = unlockhash;
+		self._unlockheight = null;
+		self.unlockheight = unlockheight;
 		self._id = null;
 		self.id = id;
+		self._is_reward = false;
+		self.is_reward = is_reward;
 	});},
 	get from_json () {return __getcm__ (this, function (cls, obj) {
 		if (arguments.length) {
@@ -1384,6 +1441,111 @@ export var MinerPayout =  __class__ ('MinerPayout', [object], {
 		}
 		self._unlockhash = value;
 	});},
+	get _get_unlockheight () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		return self._unlockheight;
+	});},
+	get _set_unlockheight () {return __get__ (this, function (self, value) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+						case 'value': var value = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (value === null) {
+			self._unlockheight = null;
+			return ;
+		}
+		if (!(isinstance (value, int))) {
+			var __except0__ = py_TypeError ("cannot assign value of type {} as a MinerPayout's unlockheight (expected: int)".format (py_typeof (value)));
+			__except0__.__cause__ = null;
+			throw __except0__;
+		}
+		if (value <= 0) {
+			var __except0__ = py_TypeError ("cannot assign value {} as a MinerPayout's unlockheight (has to be strictly positive)".format (value));
+			__except0__.__cause__ = null;
+			throw __except0__;
+		}
+		self._unlockheight = value;
+	});},
+	get _get_is_reward () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		return self._is_reward;
+	});},
+	get _set_is_reward () {return __get__ (this, function (self, value) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+						case 'value': var value = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (!(isinstance (value, bool))) {
+			var __except0__ = py_TypeError ('is is_reward is supposed to be a bool, cannot be {} ({})'.format (value, py_typeof (value)));
+			__except0__.__cause__ = null;
+			throw __except0__;
+		}
+		self._is_reward = value;
+	});},
+	get as_coin_output () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		var cond = ConditionUnlockHash (__kwargtrans__ ({unlockhash: self.unlockhash}));
+		if (self._unlockheight != null) {
+			var cond = ConditionLockTime (__kwargtrans__ ({condition: cond, lock: self._unlockheight}));
+		}
+		return CoinOutput (__kwargtrans__ ({value: self.value, condition: cond, id: self.id, is_fee: !(self._is_reward), is_reward: self._is_reward}));
+	});},
 	get _get_id () {return __get__ (this, function (self) {
 		if (arguments.length) {
 			var __ilastarg0__ = arguments.length - 1;
@@ -1423,6 +1585,8 @@ export var MinerPayout =  __class__ ('MinerPayout', [object], {
 	});}
 });
 Object.defineProperty (MinerPayout, 'id', property.call (MinerPayout, MinerPayout._get_id, MinerPayout._set_id));
+Object.defineProperty (MinerPayout, 'is_reward', property.call (MinerPayout, MinerPayout._get_is_reward, MinerPayout._set_is_reward));
+Object.defineProperty (MinerPayout, 'unlockheight', property.call (MinerPayout, MinerPayout._get_unlockheight, MinerPayout._set_unlockheight));
 Object.defineProperty (MinerPayout, 'unlockhash', property.call (MinerPayout, MinerPayout._get_unlockhash, MinerPayout._set_unlockhash));
 Object.defineProperty (MinerPayout, 'value', property.call (MinerPayout, MinerPayout._get_value, MinerPayout._set_value));;